refactor+feat: address review feedback on Sarvam provider
