+++ docs/providers/supported-providers/overview.mdx
@@ -38,7 +38,7 @@ The following table summarizes which operations are supported by each provider v
 | Replicate (`replicate/<model>`)      | ✅     | ✅   | ✅            | ✅   | ✅            | ✅        | ✅                 | ✅     | ✅              | ✅         | ✅                  | ❌              | ❌         | ❌  | ❌           | ❌  | ❌           | ✅    | ❌    | ❌           | ❌     | ❌  | ✅    | ❌          | ❌         | ❌          | ❌                   |
 | Runware (`runware/<model>`)          | ❌     | ❌   | ❌            | ❌   | ❌            | ❌        | ❌                 | ✅     | ❌              | ✅         | ❌                  | ❌              | ❌         | ❌  | ❌           | ❌  | ❌           | ❌    | ❌    | ❌           | ❌     | ❌  | ✅    | ❌          | ❌         | ❌          | ❌                   |
 | Runway (`runway/<model>`)            | ❌     | ❌   | ❌            | ❌   | ❌            | ❌        | ❌                 | ❌     | ❌              | ❌         | ❌                  | ❌              | ❌         | ❌  | ❌           | ❌  | ❌           | ❌    | ❌    | ❌           | ❌     | ❌  | ✅    | ❌          | ❌         | ❌          | ❌                   |
-| Sarvam AI (`sarvam/<model>`)         | ❌     | ❌   | ❌            | ✅   | ✅            | ✅        | ✅                 | ❌     | ❌              | ❌         | ❌                  | ❌              | ❌         | ✅  | ❌           | ✅  | ❌           | ❌    | ❌    | ❌           | ❌     | ❌  | ❌    | ❌          | ❌         | ❌          | ❌                   |
+| Sarvam AI (`sarvam/<model>`)         | ❌     | ❌   | ❌            | ✅   | ✅            | ✅        | ✅                 | ❌     | ❌              | ❌         | ❌                  | ❌              | ❌         | ✅  | ✅           | ✅  | ❌           | ❌    | ❌    | ❌           | ❌     | ❌  | ❌    | ❌          | ❌         | ❌          | ❌                   |
 | SGL (`sgl/<model>`)                  | ✅     | ✅   | ✅            | ✅   | ✅            | ✅        | ✅                 | ❌     | ❌              | ❌         | ❌                  | ❌              | ✅         | ❌  | ❌           | ❌  | ❌           | ❌    | ❌    | ❌           | ❌     | ❌  | ❌    | ❌          | ❌         | ❌          | ❌                   |
 | Vertex AI (`vertex/<model>`)         | ✅     | ❌   | ❌            | ✅   | ✅            | ✅        | ✅                 | ✅     | ❌              | ✅         | ❌                  | ❌              | ✅         | ❌  | ❌           | ❌  | ❌           | ❌    | ❌    | ✅           | ✅     | ❌  | ✅    | ❌          | ❌         | ✅          | ✅                   |
 | vLLM (`vllm/<model>`)                | ✅     | ✅   | ✅            | ✅   | ✅            | ✅        | ✅                 | ❌     | ❌              | ❌         | ❌                  | ❌              | ✅         | ❌  | ❌           | ✅  | ✅           | ❌    | ❌    | ❌           | ✅     | ❌  | ❌    | ❌          | ❌         | ❌          | ❌                   |
